Add files via upload
Write a JavaScript code to create a tip calculator for a resturant. The rules for tip are: 

a.  If the bill amount is less than 250, tip is 20%
b.  If the bill is between 250 and 500 then tip is 15% and 
c.  When the bill is above 500 then tip is 10%

Your code should accept four bill amounts from the user, calculate their tips and print the original bill as well as final bill for amount.