Add files via upload
Write a JavaScript code to divide an given array of positive integers in to two parts. First element goes to first part, second element goes to second part and third element goes to first part and so on. Now compute the sum of two parts and store in to an array and print it.

Sample Data:
---------------

Input Array: [1,3,6,2,5,10]
Output: [12,15]